requirements.txt
certifi==2018.4.16
chardet==3.0.4
ez-setup==0.9
idna==2.7
python-dotenv==0.8.2
requests==2.19.1
urllib3==1.23
wincertstore==0.2
